Implemented saved eelctions and follow feature

+++ app/src/main/java/com/example/android/politicalpreparedness/data/network/CivicsApiService.kt
@@ -34,7 +34,6 @@ private val retrofit = Retrofit.Builder()
  */
 
 interface CivicsApiService {
-    //TODO: Add elections API Call
     @GET("elections")
     suspend fun getUpcomingElections(): ElectionResponse
 
@@ -42,7 +41,7 @@ interface CivicsApiService {
     @GET("voterinfo")
     suspend fun getVoterInfoWithElection(
         @Query("address") address: String,
-        @Query("electionId") electionId: Int
+        @Query("electionId") electionId: Long
     ): VoterInfoResponse
 
     //TODO: Add representatives API Call
